Fixed bug with default constructor parameters
Fixed default constructor parameters. Reworked range data type.

+++ MHZ.cpp
@@ -26,10 +26,10 @@ const int STATUS_INCOMPLETE = -4;
 const int STATUS_NOT_READY = -5;
 const int STATUS_PWM_NOT_CONFIGURED = -6;
 const int STATUS_SERIAL_NOT_CONFIGURED = -7;
-const unsigned long RANGE_2K = 2000;
-const unsigned long RANGE_5K = 5000;
+const int RANGE_2K = 2000;
+const int RANGE_5K = 5000;
 
-MHZ::MHZ(uint8_t rxpin, uint8_t txpin, uint8_t pwmpin, uint8_t type, unsigned long range = RANGE_5K) {
+MHZ::MHZ(uint8_t rxpin, uint8_t txpin, uint8_t pwmpin, uint8_t type, uint16_t range = RANGE_5K) {
   SoftwareSerial * ss = new SoftwareSerial(rxpin, txpin);
   _pwmpin = pwmpin;
   _type = type;
@@ -49,14 +49,14 @@ MHZ::MHZ(uint8_t rxpin, uint8_t txpin, uint8_t type) {
   PwmConfigured = false;
 }
 
-MHZ::MHZ(uint8_t pwmpin, uint8_t type, unsigned long range = RANGE_5K) {
+MHZ::MHZ(uint8_t pwmpin, uint8_t type, uint16_t range = RANGE_5K) {
   _pwmpin = pwmpin;
   _type = type;
   _range = range;
   SerialConfigured = false;
 }
 
-MHZ::MHZ(Stream * serial, uint8_t pwmpin, uint8_t type, unsigned long range = RANGE_5K) {
+MHZ::MHZ(Stream * serial, uint8_t pwmpin, uint8_t type, uint16_t range = RANGE_5K) {
   _serial = serial;
   _pwmpin = pwmpin;
   _type = type;
